pyproject.toml
[tool.poetry]
name = "diplom"
version = "0.1.0"
description = ""
authors = ["Aptemond <sofartem5@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0"
python-decouple = "^3.8"
psycopg2 = "^2.9.9"
django-filter = "^23.3"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.3.0"
pillow = "^10.0.1"
djangorestframework-simplejwt = "^5.3.0"
drf-yasg = "^1.21.7"
redis = "^5.0.1"
coverage = "^7.3.4"
ruff = "^0.1.9"
twilio = "^8.11.0"
flake8 = "^6.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
